Work in progress: implementing VRML1 to SGNODE translation

+++ plugins/3d/vrml/v2/vrml2_color.cpp
@@ -207,7 +207,7 @@ bool WRL2COLOR::HasColors( void )
 
 void WRL2COLOR::GetColor( int aIndex, float& red, float& green, float& blue )
 {
-    if( aIndex < 0 || aIndex >= colors.size() )
+    if( aIndex < 0 || aIndex >= (int)colors.size() )
     {
         red = 0.8;
         green = 0.8;
